separate getIpConfig based on OS for HNS prefix changes, add subnet in UTs
